split ListLit into EmptyList and NonEmptyList
The syntax and semantics from the standard treat empty and non-empty
literals completely separately.

dhall-haskell happens to combine both at the Expr level into a single
ListLit, but we don't have to do that and it makes things simpler not
to.  It means we never have to consider an annotated, non-empty list as
special.

As a result, one of our pending parser_tests now passes for free.