fix parsing bug with reserved prefixes
This fixes the spec tests (whitespaceBuffetA.dhall was failing due to
this).

The problem was that, in IdentifierReservedPrefix, we were using the
Reserved rule which was returning the ast object for the reserved word;
rather than the text of the identifier.  The fix is to override what we
return to be c.text instead.