simplify grammar
I realised (after a discussion on dhall-lang/dhall-lang#395) that my
earlier whitespace woes have a simple solution: remove the required
whitespace from the ApplicationExpression rule.

The required whitespace is only needed for CFGs, not for PEGs, because
for CFGs optional whitespace here would introduce ambiguity.

This has the side effect that expressions like `(\(x : Natural)->x)3`
will now parse when strictly the ABNF says they shouldn't, but
dhall-haskell also successfully parses these expressions and I'm not too
worried about it.

(What were the "earlier whitespace woes"?  Well, PEG grammars are greedy
and do not backtrack.  So if you try to implement the ABNF as written,
the application-expression rule causes problems - the initial
import-expression consumes all the whitespace so there's none left for
the mandatory whitespace.  Thankfully, the PEG's greediness is also the
fix, because we can rely on it to consume the whole initial
import-expression.)