Implement CORS checking
This implements CORS checking.  This introduces a parameter to Fetch(),
the origin.  We only perform a CORS check if a remote import tries to
fetch another remote import from a different origin.

This also moves the "referential sanity" checking to Fetch() as well.
This makes more sense to me - "referential sanity" is a way to stop a
remote thing accessing a local thing (such as a file or environment
variable), CORS is another way to stop a remote thing accessing a local
thing (such as localhost or 169.254.169.254 in AWS).