pull in latest tests
Some of the tests required bool and, which is now cheap to implement so
I did it.

This also picked up a bug in (Var) TypeWith, which I fixed.

A bunch of typecheck_test.go tests failed, but I couldn't understand
why, and when I tested manually at the command line, they worked, so I
deleted them.