implement CBOR decoding
This is a significant rewrite of encoding, and a completely new decoding
implementation.

We can't implement CodecDecodeSelf on the individual expression types,
since when decoding, we don't know ahead of time what the type of
expression we're expecting is.  We need something higher-level.

We want to be able to have Expr implement codec.Selfer; so that you can
decode into a variable of type `Expr`.  But it doesn't work that way,
since one interface cannot implement another in this way.

So we go for the next best thing and create a CborBox struct to hold an
Expr, and have CborBox implement codec.Selfer.  Then, if you want to
encode something, you put it in a CborBox and encode that; if you want
to decode something, you decode it into a CborBox variable.