Fix problem with typechecking
This fixes a problem with typechecking.  The problem here was really
hard to spot, but it was because rebindLocal() had a bug for handling
NonEmptyLists of length >= 2.  The only time rebindLocal() is called is on types, and
the only types which can contain a NonEmptyList term at the moment are
equivalences.

So the minimal expression that might trigger the buggy behaviour is
something of the form

    assert : [x, y] ≡ [x, y]

because the only types that actually contain NonEmptyList values are
equivalences.