improve parser package interface
The pigeon-generated parser is a little clunky because it
returns (interface{}, error) but it's always (core.Term, error).  This
commit pulls the generated parser into parser/internal, and wraps it in
the parser package with functions that do the appropriate type
assertions.

This means we can remove a whole bunch of type assertions from elsewhere
in the code, and so can users.

This also moves the magic parser go:generate comment into the parser
package, so you now have to run `go generate ./parser` rather than `go
generate` if you're at the top level.