[unmarshal] unmarshal builtins to go funcs
I realised that Value's public interface is best captured by Callable,
which should be able to abtract over both LambdaValues and builtins.
Adding an ArgType() function to the Callable interface makes it usable
for unmarshalling purposes.

As a result, we no longer need to expose LambdaValue so it's been made
private.