Introduce {List,Optional,None}Of; hide AppValue
This introduces ListOf, OptionalOf, and NoneOf.  ListOf{x} is the new
AppValue{List, x} and similarly for the others.  Also, it renames
AppValue to appValue to unexpose it.

There are a few ideas behind this change.  First: it's easier ergonomics
to test for a Value of type ListOf than to check than a Value is an
AppValue with Fn == List.  Second: other than checking for List,
Optional and None, there's no need for consumers to introspect into an
AppValue.  A Value is either a Callable or some data (Natural, Optional,
List, RecordLit etc); all the other possible types (opValue, appValue,
localVar) are internal cases not really visible to the user.