introduce PlainTextLit value type
This splits core.TextLit into interpolatedText and PlainTextLit.  The
idea is that, like most other compound expressions (appValue, opValue
etc), users won't have to deal with interpolatedText directly.  The only
kind of Text they have to deal with now is a fully-evaluated plain text
literal with no interpolations.

This makes the public interface simpler at the cost of a couple of extra
cases in switch statements here and there.