Use `dhall` struct tag instead of `json` (#15)
I originally used `json` to try to piggyback on existing intrumented
code.  However, this is problematic if a user wants to be able to decode
as Dhall in one way but serialize as JSON in another way.

Therefore, this commit switches us to use the `dhall` tag.

Note also that `json` tags support extra metadata after a comma; we
never used that metadata but we had code to remove it.  Also `json`
supports ignoring a field completely with a tag of `"-"`.  This commit
removes all that code; `dhall` tags are just simple strings.

It's possible that in future we will still want to piggyback on `json`
tags, or add extra metadata to `dhall` tags, but for now I want to keep
dhall-golang as simple as possible and not add features until they are
demonstrably required.

Resolves #10.