[unmarshal] refactor decode()
This refactors the decode() switch to go through the Kinds in the order
given in the reflect docs.  It also adds a panic() at the bottom and
returns early from all the successful cases.

This will allow us to change decode() to return an error instead of
panicking in future.