Fix Natural/Integer uint/int behaviour
I confused myself with this, so it's worth writing down the intention:

 - Natural should be encodable/decodable to/from uint types
 - Integer should be encodable/decodable to/from int types
 - Natural should be decodable to int (but not encodable from int)

The code was broken because it allowed encoding Natural from int which
we want to disallow to prevent a go user passing a negative number to a
decoded Dhall function.

Decoding Natural to int types is safe and seems like a good convenience
feature to have.

In principle, we could also allow encoding uint types to Integer but I
see much less value here:

 - Integer is less commonly used than Natural
 - uint is less commonly used than int
 - this only applies to Dhall function parameters, not to plain values

so for now I have decided not to implement this.