introduce unionVal, unionConstructor Value types
It's a pain to deal with union values and constructors in terms of their
syntax.  This introduces unionVal and unionConstructor so that Value
represents semantic union values, not syntactic field accesses of union
types.

This makes the code for evaluating Merge terms more robust - only union
values can trigger it, not other field accesses.  (There was a bug here
before this commit.)