Perform alpha normalization at quote time
This moves the alpha normalization step from eval time to quote time.
This makes sense because it's just as easy to do at either, but doing it
at quote time allows us to Eval() a Term just once and get its normal
form and alpha-beta-normal form by calling Quote() and
QuoteAlphaNormal() respectively on it.

This is particularly helpful for when we want the semantic hash of an
expression but don't necessarily want to lose all labels.