Fix Text/replace builtin when the replacement is interpolated text
Previously, `Text/replace "foo" "bar${x}baz" "foofoo"` did not work as
expected.  This fixes it by extracting the code from eval.go that
deals with evaluating Text into a textValBuilder struct, and then
reusing that textValBuilder in the Text/replace implementation.

This pins us to a dhall-lang commit just after
https://github.com/dhall-lang/dhall-lang/pull/1084 was merged; this is
the same as verison 19.0.0 with a fix for
TestNormalization/unit/TextReplaceAbstractA.